Milestone14

+++ backend/src/controller/Product.Controller.js
@@ -69,14 +69,18 @@ const updateProductController = async (req, res) => {
     const { title, description, discountedPrice, originalPrice, quantity, category, rating } = req.body
     const { id } = req.params
     try {
-        
-        const checkifproductExists = await ProductModel.findOne({_id:id})
+
+
+        const checkifproductExists = await ProductModel.findOne({ _id: id })
+
 
         if (!checkifproductExists) {
             return res.status(404).send({ message: 'Product not found' })
         }
 
-        const ImgArray = req.files && req.files.map(async(singleFile, index) => {
+
+        const ImgArray = req.files && req.files.map(async (singleFile, index) => {
+
             return cloudinary.uploader.upload(singleFile.path, {
                 folder: 'uploads'
             })
@@ -89,8 +93,10 @@ const updateProductController = async (req, res) => {
         const Imagedata = req.files && (await Promise.all(ImgArray))
         const UpdatedImages = req.files ? Imagedata : req.body.images;
         const findAndupdate = await ProductModel.findByIdAndUpdate(
-            { _id: id }, 
-            { title, description, discountedPrice, originalPrice, quantity, category, rating, images: UpdatedImages }, 
+
+            { _id: id },
+            { title, description, discountedPrice, originalPrice, quantity, category, rating, images: UpdatedImages },
+
             { new: true }
         )
 
@@ -99,6 +105,8 @@ const updateProductController = async (req, res) => {
         res.status(500).send({ message: err.message, success: false })
     }
 }
+
+
 
 const getSingleProductDocumentController=async(req,res)=>{
     const {id}=req.params
@@ -116,6 +124,7 @@ const getSingleProductDocumentController=async(req,res)=>{
 }
 
 const deleteSingleProductController = async (req, res) => {
+
     const { id } = req.params
 
     try {
@@ -125,14 +134,32 @@ const deleteSingleProductController = async (req, res) => {
             return res.status(404).send({ message: 'Product not found' })
         }
 
-        await ProductModel.findByIdAndDelete({_id:id})
-        const newData=await ProductModel.find()
-        return res.status(200).send({ message: "Product Fetched Successfully", newDatadata, success: true })
+        return res.status(200).send({ message: "Product Fetched Successfully", data, success: true })
     } catch (err) {
-        return res.status(200).send({ message: err.message,success:false })
+        return res.status(200).send({ message: err.message })
     }
 }
 
+const deleteSingleProductController = async (req, res) => {
+    const { id } = req.params
+
+    try {
+        const data = await ProductModel.findOne({ _id: id })
+
+        if (!data) {
+            return res.status(404).send({ message: 'Product not found' })
+        }
+
+        await ProductModel.findByIdAndDelete({ _id: id })
+        const newData = await ProductModel.find()
+        return res.status(200).send({ message: "Product Fetched Successfully", newDatadata, success: true })
+    } catch (err) {
+        return res.status(200).send({ message: err.message, success: false })
+    }
+}
+
+
 module.exports = { createProductController, getProductDataController,updateProductController,getSingleProductDocumentController,deleteSingleProductController }
+
